I think I finished with the parser and everything except the visitors :D
I had to change the grammar so we could only work with
uint, ulong, int and long. For that reason, I had to
change the tokens, scanner, parser and exp. Also, we do
not use strings in this project but this compiler could
scan and parse them in a basic way...

Also, I change the class number exp so the value could
be large enough to store long or ulong values so I changed
it to unsigned long long int the attribute value in class
NumberExp

I am nearly finished with this project and with some help in the codegen
I think I can finish ^-^

Now, I also change one of my own test cases and I realized some stuff like
the string, NumberExp, etc. to change in my code and probably in other
test cases too, I am going to do that later or tomorrow xD