So basically I change the grammar because the rule parseStmtList was not correct.
I modify the scanner because it did not recognize the
number, only the suffix, when I use a value like
1U, I did not scan the number but the suffix yes.
So I change that it scan both of them. I also change
the condition of parser Factor ::= num for that reason
but it was a minor change (just change the parenthesis
in one of the if conditions)

I modify exp files so number exp use long long value
instead of unsigned long long. For that reason, I modify
the parser too and clarify the functionality of stmt list

All of these errors I had seem them in some of the test Test_Cases
that I modified and they can be scanned and parse. However,
I am still not finish with the test cases...